fix(routes/whitehouse): new theme
